Update Script & GitIgnore
Git Ignore: Tif files used for testing are not being uploaded at present.
Script: Used script from prac4 to get the full image to display correctly. Code is largely copied and not tidied or personalised so that will come next. Then we will define a function to calculcate NDVI based on the correct bands.