Update NDVI_script.ipynb
There now exists a large function to make the NDVI and classify it. If time allows I will try to break this into two smaller functions, but at present it works and gives the desired outcome.
The stacked image plots, but is a bit dark. Might try edit Bob's code with min/max values.

Considering adding function for ndmi like ndvi, and then a method of exporting the result.